fix/#140: 도메인 모델에 따른 UpcomingSessionSection 컴포넌트로 변경
